refactor: renamed first row component
